+++ sample/application/controllers/Test.php
@@ -8,61 +8,51 @@ use \X\Util\ImageHelper;
 class Test extends AppController {
   protected $model = 'UserModel';
 
-  public function password_hash_test() {
-    $password = 'password';
-    $passwordHash = Cipher::encode_sha256($password);
-    Logger::display('$password=', $password);
-    Logger::display('$passwordHash=', $passwordHash);
-  }
-  
-  public function image_writing_test() {
-    $filePath = FCPATH . 'upload/test.png';
-    $dataUrl = 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAGQAAABkCAIAAAD/gAIDAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAAAAJcEhZcwAAEnQAABJ0Ad5mH3gAAAA0SURBVHhe7cEBDQAAAMKg909tDjcgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAIAbNXWUAAEE/b5iAAAAAElFTkSuQmCC';
-    ImageHelper::putBase64($dataUrl, $filePath);
-    Logger::display("Write {$filePath}");
-  }
+  /**
+   * @example php public/index.php test/form_validation
+   */
+  public function form_validation() {
+    // // Check the host name.
+    // $valid = $this->form_validation
+    //   ->set_data([
+    //     'hostname1' => 'example.com',
+    //     'hostname2' => 'localhost',
+    //     'hostname3' => 'c-61-123-45-67.hsd1.co.comcast.net',
+    //     'hostname4' => 'example',
+    //   ])
+    //   ->set_rules('hostname1', 'hostname1', 'required|hostname')
+    //   ->set_rules('hostname2', 'hostname2', 'required|hostname')
+    //   ->set_rules('hostname3', 'hostname3', 'required|hostname')
+    //   ->set_rules('hostname4', 'hostname4', 'required|hostname')
+    //   ->run();
+    // Logger::display('Host name check result=', var_export($valid, true));
+    // Logger::display('Host name input error:', $this->form_validation->error_array());
 
-  // public function error() {
-  //   throw new \RuntimeException();
-  // }
+    // // Reset form validation.
+    // $this->form_validation->reset_validation();
 
-  public function form_validation_test() {
-    $data = [
-      'hostname1' => 'example.com',
-      'hostname2' => 'localhost',
-      'hostname3' => 'c-61-123-45-67.hsd1.co.comcast.net',
-      'hostname4' => 'example',
-    ];
-    $isValid = $this->form_validation
-      ->set_data($data)
-      ->set_rules('hostname1', 'hostname1', 'required|hostname')
-      ->set_rules('hostname2', 'hostname2', 'required|hostname')
-      ->set_rules('hostname3', 'hostname3', 'required|hostname')
-      ->set_rules('hostname4', 'hostname4', 'required|hostname')
+    // Check the file path.
+    $valid = $this->form_validation
+      ->set_data([
+        'path1' => '/usr/lib',// It should be valid.
+        'path2' => 'usr/lib',// It should be valid.
+        'path3' => '/usr/lib',// It should be invalid.
+        'path4' => 'usr/lib',// It should be valid.
+      ])
+      ->set_rules('path1', 'path1', 'is_path')
+      ->set_rules('path2', 'path2', 'is_path')
+      ->set_rules('path3', 'path3', 'is_path[true]')
+      ->set_rules('path4', 'path4', 'is_path[true]')
       ->run();
-    Logger::display('$isValid=', $isValid ? 1 : 0);
-    $error = $this->form_validation->error_array();
-    Logger::display('$error=', $error);
+    Logger::display('Path check result=', var_export($valid, true));
+    if (!$valid)
+      Logger::display('Path input error:', $this->form_validation->error_array());
   }
 
   /**
-   * ```sh
-   * CI_ENV=development php public/index.php test/log_test
-   * ```
+   * @example CI_ENV=development php public/index.php test/sanitize_sql
    */
-  public function log_test() {
-    Logger::debug('debug');
-    Logger::info('info');
-    Logger::error('error');
-    Logger::display('display');
-  }
-
-  /**
-   * ```sh
-   * CI_ENV=development php public/index.php test/sanitize_sql_parameters
-   * ```
-   */
-  public function sanitize_sql_parameters() {
+  public function sanitize_sql() {
     $this->UserModel
       ->where('id', 1)
       ->get()
@@ -85,10 +75,5 @@ class Test extends AppController {
     $sql = $this->UserModel->last_query();
     Logger::display($sql);
     // => SELECT * FROM `user` WHERE id =  '1' OR id=2
-  }
-
-  public function get_ip() {
-    $ip = $this->input->ip_address();
-    Logger::display($ip);
   }
 }